requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cloudinary==1.33.0
Django==4.2.1
django-ckeditor==6.5.1
django-decouple==2.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
psycopg2==2.9.6
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
